Move case notes to a standalone service
OffenderService is overloaded, this commit moves the case notes logic into a standalone service.

I’ve explicitly added a supress warning that was previously covered by the baseline detekt ignore file
