+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/unit/service/OffenderServiceTest.kt
@@ -15,59 +15,37 @@ import org.junit.jupiter.params.provider.CsvSource
 import org.springframework.http.HttpMethod
 import org.springframework.http.HttpStatus
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.ApDeliusContextApiClient
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.CaseNotesClient
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.ClientResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.ClientResult.Failure.StatusCode
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.PrisonerAlertsApiClient
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.PrisonsApiClient
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.config.ExcludedCategoryBindingModel
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.config.PrisonAdjudicationsConfigBindingModel
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.config.PrisonCaseNotesConfigBindingModel
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.datasource.OffenderDetailsDataSource
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.AdjudicationFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.AdjudicationsPageFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.AgencyFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.CaseAccessFactory
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.CaseNoteFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.CaseSummaryFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.OffenderDetailsSummaryFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.PersonSummaryInfoResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.deliuscontext.CaseSummaries
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.deliuscontext.UserAccess
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.deliuscontext.UserOffenderAccess
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.prisonsapi.CaseNotesPage
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.problem.InternalServerErrorProblem
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.AuthorisableActionResult
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.CasResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.LaoStrategy
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.LaoStrategy.CheckUserAccess
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.OffenderService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.unit.util.ObjectMapperFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.ClientResultFailureArgumentsProvider
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.randomNumberChars
-import java.time.LocalDate
 
 class OffenderServiceTest {
   private val mockPrisonsApiClient = mockk<PrisonsApiClient>()
   private val mockPrisonerAlertsApiClient = mockk<PrisonerAlertsApiClient>()
-  private val mockCaseNotesClient = mockk<CaseNotesClient>()
   private val mockApDeliusContextApiClient = mockk<ApDeliusContextApiClient>()
   private val mockOffenderDetailsDataSource = mockk<OffenderDetailsDataSource>()
 
-  private val prisonCaseNotesConfigBindingModel = PrisonCaseNotesConfigBindingModel().apply {
-    lookbackDays = 30
-    prisonApiPageSize = 2
-    excludedCategories = listOf(
-      ExcludedCategoryBindingModel().apply {
-        this.category = "CATEGORY"
-        this.subcategory = "EXCLUDED_SUBTYPE"
-      },
-      ExcludedCategoryBindingModel().apply {
-        this.category = "EXCLUDED_CATEGORY"
-        this.subcategory = null
-      },
-    )
-  }
   private val adjudicationsConfigBindingModel = PrisonAdjudicationsConfigBindingModel().apply {
     prisonApiPageSize = 2
   }
@@ -75,10 +53,8 @@ class OffenderServiceTest {
   private val offenderService = OffenderService(
     mockPrisonsApiClient,
     mockPrisonerAlertsApiClient,
-    mockCaseNotesClient,
     mockApDeliusContextApiClient,
     mockOffenderDetailsDataSource,
-    prisonCaseNotesConfigBindingModel,
     adjudicationsConfigBindingModel,
   )
 
@@ -446,154 +422,6 @@ class OffenderServiceTest {
 
       assertThrows<Throwable> { offenderService.canAccessOffenders("distinguished.name", listOf(crn)) }
     }
-  }
-
-  @Test
-  fun `getFilteredPrisonCaseNotesByNomsNumber returns NotFound when Case Notes request returns a 404`() {
-    val nomsNumber = "NOMS456"
-
-    every {
-      mockCaseNotesClient.getCaseNotesPage(
-        nomsNumber = nomsNumber,
-        from = LocalDate.now().minusDays(prisonCaseNotesConfigBindingModel.lookbackDays!!.toLong()),
-        page = 0,
-        pageSize = 2,
-      )
-    } returns StatusCode(HttpMethod.GET, "/api/offenders/$nomsNumber/v2", HttpStatus.NOT_FOUND, null)
-
-    var result = offenderService.getFilteredPrisonCaseNotesByNomsNumber(nomsNumber, false)
-    assertThat(result).isInstanceOf(CasResult.NotFound::class.java)
-  }
-
-  @Test
-  fun `getFilteredPrisonCaseNotesByNomsNumber returns Unauthorised when Case Notes request returns a 403`() {
-    val nomsNumber = "NOMS456"
-
-    every {
-      mockCaseNotesClient.getCaseNotesPage(
-        nomsNumber = nomsNumber,
-        from = LocalDate.now().minusDays(prisonCaseNotesConfigBindingModel.lookbackDays!!.toLong()),
-        page = 0,
-        pageSize = 2,
-      )
-    } returns StatusCode(HttpMethod.GET, "/api/offenders/$nomsNumber/v2", HttpStatus.FORBIDDEN, null)
-
-    var result = offenderService.getFilteredPrisonCaseNotesByNomsNumber(nomsNumber, false)
-    assertThat(result).isInstanceOf(CasResult.Unauthorised::class.java)
-  }
-
-  @Test
-  fun `getFilteredPrisonCaseNotesByNomsNumber returns Success, traverses pages from Client & excludes categories + subcategories`() {
-    val nomsNumber = "NOMS456"
-
-    val caseNotesPageOne = listOf(
-      CaseNoteFactory().produce(),
-      CaseNoteFactory().produce(),
-      CaseNoteFactory().withType("EXCLUDED_TYPE").produce(),
-    )
-
-    val caseNotesPageTwo = listOf(
-      CaseNoteFactory().produce(),
-      CaseNoteFactory().produce(),
-      CaseNoteFactory().withType("TYPE").withSubType("EXCLUDED_SUBTYPE").produce(),
-    )
-
-    every {
-      mockCaseNotesClient.getCaseNotesPage(
-        nomsNumber = nomsNumber,
-        from = LocalDate.now().minusDays(prisonCaseNotesConfigBindingModel.lookbackDays!!.toLong()),
-        page = 0,
-        pageSize = 2,
-      )
-    } returns ClientResult.Success(
-      HttpStatus.OK,
-      CaseNotesPage(
-        totalElements = 6,
-        totalPages = 2,
-        number = 1,
-        content = caseNotesPageOne,
-      ),
-    )
-
-    every {
-      mockCaseNotesClient.getCaseNotesPage(
-        nomsNumber = nomsNumber,
-        from = LocalDate.now().minusDays(prisonCaseNotesConfigBindingModel.lookbackDays!!.toLong()),
-        page = 1,
-        pageSize = 2,
-      )
-    } returns ClientResult.Success(
-      HttpStatus.OK,
-      CaseNotesPage(
-        totalElements = 4,
-        totalPages = 2,
-        number = 2,
-        content = caseNotesPageTwo,
-      ),
-    )
-
-    val result = offenderService.getFilteredPrisonCaseNotesByNomsNumber(nomsNumber, false)
-
-    assertThat(result).isInstanceOf(CasResult.Success::class.java)
-    result as CasResult.Success
-    assertThat(result.value).containsAll(caseNotesPageOne.subList(0, 1) + caseNotesPageTwo.subList(0, 1))
-  }
-
-  @Test
-  fun `getFilteredPrisonCaseNotesByNomsNumber returns specified case note types when getSpecificNoteTypes is true`() {
-    val nomsNumber = "NOMS456"
-
-    val caseNotesPageOne = listOf(
-      CaseNoteFactory().produce(),
-      CaseNoteFactory().produce(),
-      CaseNoteFactory().withTypeDescription(null).withType("Enforcement").produce(),
-    )
-
-    val caseNotesPageTwo = listOf(
-      CaseNoteFactory().produce(),
-      CaseNoteFactory().produce(),
-      CaseNoteFactory().withTypeDescription("Alert").produce(),
-    )
-
-    every {
-      mockCaseNotesClient.getCaseNotesPage(
-        nomsNumber = nomsNumber,
-        from = LocalDate.now().minusDays(prisonCaseNotesConfigBindingModel.lookbackDays!!.toLong()),
-        page = 0,
-        pageSize = 2,
-      )
-    } returns ClientResult.Success(
-      HttpStatus.OK,
-      CaseNotesPage(
-        totalElements = 6,
-        totalPages = 2,
-        number = 1,
-        content = caseNotesPageOne,
-      ),
-    )
-
-    every {
-      mockCaseNotesClient.getCaseNotesPage(
-        nomsNumber = nomsNumber,
-        from = LocalDate.now().minusDays(prisonCaseNotesConfigBindingModel.lookbackDays!!.toLong()),
-        page = 1,
-        pageSize = 2,
-      )
-    } returns ClientResult.Success(
-      HttpStatus.OK,
-      CaseNotesPage(
-        totalElements = 4,
-        totalPages = 2,
-        number = 2,
-        content = caseNotesPageTwo,
-      ),
-    )
-
-    val result = offenderService.getFilteredPrisonCaseNotesByNomsNumber(nomsNumber, true)
-
-    assertThat(result).isInstanceOf(CasResult.Success::class.java)
-    result as CasResult.Success
-    assertThat(result.value).containsAll(caseNotesPageTwo.subList(2, 3) + caseNotesPageTwo.subList(2, 3))
   }
 
   @Test
